+++ runtime-kotlin/runtime-ports/src/main/kotlin/skillbill/ports/workflow/gitops/model/ReadinessTreeIdentity.kt
@@ -1,0 +1,7 @@
+package skillbill.ports.workflow.gitops.model
+
+data class ReadinessTreeIdentity(
+  val sourceTreeSha: String,
+  val baseRefSha: String,
+  val headSha: String,
+)